`split_closure: `Improve `unsafe`-correctness + minor improvements
`split_closure` does not need to be marked `unsafe`: one cannot cause unsoundness with the obtained raw pointer and `unsafe fn` function pointer, unless further `unsafe` is used.

Hence removing the `unsafe`-ness (to call) of `split_closure`.

That being said, the pattern `unsafe { unsafe_fn_ptr(void_ptr, <args...) }` is, on the other hand, expected to be sound, at least if done within the lifetime of the borrow over `C`.

This is "problematic" because `Split` itself is not marked `unsafe`, so once could go and write a troll / evil implementation of `Split`, with which the aformentioned pattenr would be unsound.

The obvious solution would be to have `Split` become an `unsafe trait`, but after thinking about it, I didn't like that approach: given the lack of variadic generics, it is almost impossible to explicitly name / describe within the type system that the `Trampoline` is expected to be some `unsafe fn` pointer, which ought to be extracted out of an `FnMut` closure, and which must be sound to call with an unaliased raw pointer to that closure. Because of that, the other approach is to simply `Seal` the trait within the privacy boundary, so that all the logic related to the trait is guaranteed to be self-contained within this one file.

___

Also, I've removed the unneeded `Ret` param of `Split`, and made obtaining the `trampoline` `const`.