Write backtrace attached anyhow::Error
Commit #c306f0cc2f8 replaced failure::Error by anyhow::Error but didn't
change the string conversion (which is to use ToString implementation).

The common approach to print anyhow::Error is to either use the Debug
formatter or to use the alternate selector "{:#}" to also print the
causes attached to it.